pyproject.toml
[project]
name = "UserIdTgBot"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "vadyushkins", email = "vadim.i.abzalov@gmail.com"},
]
dependencies = [
    "python-telegram-bot>=21.5",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
